Move LocalFeedItem to its own file
